Refactor Parquet RleDecoder
This commit contains the following changes:

1) Extract common operations from RleDecoder to the newly added
RleBpDecoder class, and change repeatDecoder_ and defineDecoder_ to be
of RleBpDecoder instead of RleDecoder. The latter one was too complex
to serve the purpose.

2) Change RleDecoder to ParquetDictionaryIdDecoder, and make it extends
the RleBpDecoder class instead of IntDecoder. This new class
ParquetDictionaryIdDecoder only handles dictionary Ids which is
RLE/BitPack encoded in Parquet.

3) Extract bit unpacking related code from IntDecoder to BitUnpacker.
The code in this class  will be further optimized in future commits.
