+++ velox/dwio/common/tests/CMakeLists.txt
@@ -17,9 +17,9 @@ add_subdirectory(utils)
 add_executable(
   velox_dwio_common_test
   BitConcatenationTest.cpp
-  DecodeBitsTest.cpp
   ChainedBufferTests.cpp
   DataBufferTests.cpp
+  DecodeBitsTest.cpp
   DecoderUtilTest.cpp
   LoggedExceptionTest.cpp
   RangeTests.cpp
@@ -66,3 +66,8 @@ target_link_libraries(
   ${ZSTD}
   ${ZLIB_LIBRARIES}
   ${TEST_LINK_LIBS})
+
+add_executable(velox_dwio_common_bitpack_decoder_benchmark
+               BitPackDecoderBenchmark.cpp)
+target_link_libraries(velox_dwio_common_bitpack_decoder_benchmark
+                      velox_dwio_common ${FOLLY} ${FOLLY_BENCHMARK})
