+++ velox/dwio/common/tests/IntDecoderBenchmark.cpp
@@ -48,17 +48,6 @@ std::vector<uint64_t> randomInts_u64;
 std::vector<uint64_t> randomInts_u64_result;
 std::vector<char> buffer_u64;
 
-// Array of bit packed representations of randomInts_u32. The array at index i
-// is packed i bits wide and the values come from the low bits of
-std::vector<std::vector<uint64_t>> bitPackedData;
-
-std::vector<uint32_t> result32;
-
-std::vector<int32_t> allRowNumbers;
-std::vector<int32_t> oddRowNumbers;
-RowSet allRows;
-RowSet oddRows;
-
 uint64_t readVuLong(const char* buffer, size_t& len) {
   if (LIKELY(len >= folly::kMaxVarintLength64)) {
     const char* p = buffer;
@@ -949,144 +938,6 @@ BENCHMARK_RELATIVE(decodeNew_64) {
       randomInts_u64.size(), buffer_u64.data(), randomInts_u64_result.data());
 }
 
-// Naive unpacking, original version of IntDecoder::decodeBitsLE.
-template <typename T>
-void naiveDecodeBitsLE(
-    const uint64_t* FOLLY_NONNULL bits,
-    int32_t bitOffset,
-    RowSet rows,
-    int32_t rowBias,
-    uint8_t bitWidth,
-    const char* bufferEnd,
-    T* FOLLY_NONNULL result) {
-  uint64_t mask = bits::lowMask(bitWidth);
-  auto numRows = rows.size();
-  if (bitWidth > 56) {
-    for (auto i = 0; i < numRows; ++i) {
-      auto bit = bitOffset + (rows[i] - rowBias) * bitWidth;
-      result[i] = bits::detail::loadBits<T>(bits, bit, bitWidth) & mask;
-    }
-    return;
-  }
-  auto FOLLY_NONNULL lastSafe = bufferEnd - sizeof(uint64_t);
-  int32_t numSafeRows = numRows;
-  bool anyUnsafe = false;
-  if (bufferEnd) {
-    const char* endByte = reinterpret_cast<const char*>(bits) +
-        bits::roundUp(bitOffset + (rows.back() - rowBias + 1) * bitWidth, 8) /
-            8;
-    // redzone is the number of bytes at the end of the accessed range that
-    // could overflow the buffer if accessed 64 its wide.
-    int64_t redZone =
-        sizeof(uint64_t) - static_cast<int64_t>(bufferEnd - endByte);
-    if (redZone > 0) {
-      anyUnsafe = true;
-      auto numRed = (redZone + 1) * 8 / bitWidth;
-      int32_t lastSafeIndex = rows.back() - numRed;
-      --numSafeRows;
-      for (; numSafeRows >= 1; --numSafeRows) {
-        if (rows[numSafeRows - 1] < lastSafeIndex) {
-          break;
-        }
-      }
-    }
-  }
-  for (auto i = 0; i < numSafeRows; ++i) {
-    auto bit = bitOffset + (rows[i] - rowBias) * bitWidth;
-    auto byte = bit / 8;
-    auto shift = bit & 7;
-    result[i] = (*reinterpret_cast<const uint64_t*>(
-                     reinterpret_cast<const char*>(bits) + byte) >>
-                 shift) &
-        mask;
-  }
-  if (anyUnsafe) {
-    auto lastSafeWord = bufferEnd - sizeof(uint64_t);
-    VELOX_DCHECK(lastSafeWord);
-    for (auto i = numSafeRows; i < numRows; ++i) {
-      auto bit = bitOffset + (rows[i] - rowBias) * bitWidth;
-      auto byte = bit / 8;
-      auto shift = bit & 7;
-      result[i] = IntDecoder<false>::safeLoadBits(
-                      reinterpret_cast<const char*>(bits) + byte,
-                      shift,
-                      bitWidth,
-                      lastSafeWord) &
-          mask;
-    }
-  }
-}
-
-template <typename T>
-void unpackNaive(RowSet rows, uint8_t bitWidth, T* result) {
-  auto data = bitPackedData[bitWidth].data();
-  auto numBytes = bits::roundUp((rows.back() + 1) * bitWidth, 8) / 8;
-  auto end = reinterpret_cast<const char*>(data) + numBytes;
-  naiveDecodeBitsLE(data, 0, rows, 0, bitWidth, end, result32.data());
-}
-
-template <typename T>
-void unpackFast(RowSet rows, uint8_t bitWidth, T* result) {
-  auto data = bitPackedData[bitWidth].data();
-  auto numBytes = bits::roundUp((rows.back() + 1) * bitWidth, 8) / 8;
-  auto end = reinterpret_cast<const char*>(data) + numBytes;
-  IntDecoder<false>::decodeBitsLE(
-      data,
-      0,
-      rows,
-      0,
-      bitWidth,
-      end,
-      reinterpret_cast<int32_t*>(result32.data()));
-}
-
-#define BIT_BM_CASE_32(width)                       \
-  BENCHMARK(unpackNaive##width##_32) {              \
-    unpackNaive(allRows, width, result32.data());   \
-  }                                                 \
-                                                    \
-  BENCHMARK_RELATIVE(unpackFast##width##_32) {      \
-    unpackFast(allRows, width, result32.data());    \
-  }                                                 \
-                                                    \
-  BENCHMARK_RELATIVE(unpackNaive##width##_32_odd) { \
-    unpackNaive(oddRows, width, result32.data());   \
-  }                                                 \
-                                                    \
-  BENCHMARK_RELATIVE(unpackFast##width##_32_odd) {  \
-    unpackFast(oddRows, 7, result32.data());        \
-  }
-
-BIT_BM_CASE_32(7)
-BIT_BM_CASE_32(8)
-BIT_BM_CASE_32(13)
-BIT_BM_CASE_32(16)
-BIT_BM_CASE_32(22)
-BIT_BM_CASE_32(24)
-BIT_BM_CASE_32(31)
-
-void populateBitPacked() {
-  bitPackedData.resize(32);
-  for (auto bitWidth = 2; bitWidth < 32; ++bitWidth) {
-    auto numWords = bits::roundUp(randomInts_u32.size() * bitWidth, 64) / 64;
-    bitPackedData[bitWidth].resize(numWords);
-    auto source = reinterpret_cast<uint64_t*>(randomInts_u32.data());
-    auto destination =
-        reinterpret_cast<uint64_t*>(bitPackedData[bitWidth].data());
-    for (auto i = 0; i < randomInts_u32.size(); ++i) {
-      bits::copyBits(source, i * 32, destination, i * bitWidth, bitWidth);
-    }
-  }
-  allRowNumbers.resize(randomInts_u32.size());
-  std::iota(allRowNumbers.begin(), allRowNumbers.end(), 0);
-  oddRowNumbers.resize(randomInts_u32.size() / 2);
-  for (auto i = 0; i < oddRowNumbers.size(); i++) {
-    oddRowNumbers[i] = i * 2 + 1;
-  }
-  allRows = RowSet(allRowNumbers);
-  oddRows = RowSet(oddRowNumbers);
-}
-
 int32_t main(int32_t argc, char* argv[]) {
   folly::init(&argc, &argv);
 
@@ -1120,8 +971,6 @@ int32_t main(int32_t argc, char* argv[]) {
     randomInts_u64.push_back(randomInt);
     pos = writeVulongToBuffer(randomInt, buffer_u64.data(), pos);
   }
-  populateBitPacked();
-  result32.resize(randomInts_u32.size());
 
   randomInts_u64_result.resize(randomInts_u64.size());
   len_u64 = pos;
